MainWindow:Open dialogue can be opened with a shortcut (cmd-o)
This commmit adds a shortcut so the open connection dialogue can be opened with cmd-O.
This is a pain in the Windows world as cmd(win) button is used for system events,
and not per application. Instead ctrl should be used. But as I am, presently,
the sole user of this fork I choose to not solve the problem. Right now.

This commit also introduced the graphical cue that text in italic is a shortcut.
Again - the standard is different in Windows with an underlined letter.
But then - again - I choose to not solve the problem. Now.

This commit solves #6